Extract tests for isInfinity to own file
